Crate hibitset

hibitset

Provides hierarchical bit sets, which allow very fast iteration on sparse data structures.

Structs

AtomicBitSet
This is similar to a BitSet but allows setting of value without unique ownership of the structure

BitIter
An Iterator over a BitSetLike structure.

BitParIter
A ParallelIterator over a BitSetLike structure.

BitProducer
Allows splitting and internally iterating through BitSet.

BitSet
A BitSet is a simple set designed to track entity indices for which a certain component exists. It does not track the Generation of the entities that it contains.

BitSetAnd
BitSetAnd takes two BitSetLike items, and merges the masks returning a new virtual set, which represents an intersection of the two original sets.

BitSetNot
BitSetNot takes a BitSetLike item, and produced an inverted virtual set. Note: the implementation is sub-optimal because layers 1-3 are not active.

BitSetOr
BitSetOr takes two BitSetLike items, and merges the masks returning a new virtual set, which represents an merged of the two original sets.

BitSetXor
BitSetXor takes two BitSetLike items, and merges the masks returning a new virtual set, which represents an merged of the two original sets.

Traits

BitSetLike
A generic interface for BitSetLike-like types.