Struct hibitset::BitSetOr

pub struct BitSetOr<A: BitSetLike, B: BitSetLike>(pub A, pub B);

BitSetOr takes two BitSetLike items, and merges the masks returning a new virtual set, which represents an merged of the two original sets.